refactor: remove SharedFilterProvider
