+++ packages/core/client/src/block-provider/BlockProvider.tsx
@@ -19,7 +19,6 @@ import { ACLCollectionProvider } from '../acl/ACLProvider';
 import { CollectionProvider, useCollection, useCollectionManager } from '../collection-manager';
 import { FilterBlockRecord } from '../filter-provider/FilterProvider';
 import { useRecordIndex } from '../record-provider';
-import { SharedFilterProvider } from './SharedFilterProvider';
 import { useTemplateBlockContext } from './TemplateBlockProvider';
 import { useAssociationNames } from './hooks';
 
@@ -321,11 +320,9 @@ export const BlockProvider = (props: {
         <BlockAssociationContext.Provider value={association}>
           <BlockResourceContext.Provider value={resource}>
             <BlockRequestProvider {...props} updateAssociationValues={updateAssociationValues} params={params}>
-              <SharedFilterProvider {...props} params={params}>
-                <FilterBlockRecord {...props} params={params}>
-                  {props.children}
-                </FilterBlockRecord>
-              </SharedFilterProvider>
+              <FilterBlockRecord {...props} params={params}>
+                {props.children}
+              </FilterBlockRecord>
             </BlockRequestProvider>
           </BlockResourceContext.Provider>
         </BlockAssociationContext.Provider>
