+++ packages/core/client/src/block-provider/index.tsx
@@ -2,10 +2,8 @@ export * from './BlockProvider';
 export * from './BlockSchemaComponentProvider';
 export * from './FilterFormBlockProvider';
 export * from './FormBlockProvider';
-export * from './SharedFilterProvider';
+export * from './FormFieldProvider';
 export * from './TableBlockProvider';
 export * from './TableFieldProvider';
 export * from './TableSelectorProvider';
-export * from './FormFieldProvider';
-export * from './SharedFilterProvider';
 export * from './hooks';
