+++ packages/core/client/src/schema-settings/DataTemplates/FormDataTemplates.tsx
@@ -5,8 +5,8 @@ import { Tree as AntdTree } from 'antd';
 import _ from 'lodash';
 import React, { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import { mergeFilter } from '../../block-provider';
 import { useCollectionManager } from '../../collection-manager';
+import { mergeFilter } from '../../filter-provider/utils';
 import { SchemaComponent, SchemaComponentContext, removeNullCondition } from '../../schema-component';
 import { ITemplate } from '../../schema-component/antd/form-v2/Templates';
 import { VariableInput } from '../VariableInput';
